requirements.txt
SQLAlchemy
mysql-connector-python
python-dotenv
tabulate
